requirements.txt
altgraph==0.16.1
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Django==2.2.6
et-xmlfile==1.0.1
Flask==1.1.1
future==0.17.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
line-bot-sdk==1.14.0
macholib==1.11
MarkupSafe==1.1.1
openpyxl==2.6.2
pefile==2019.4.18
pipenv==2018.11.26
psycopg2==2.8.3
PyInstaller==3.4
pytz==2019.3
requests==2.22.0
selenium==3.141.0
soupsieve==1.9.4
sqlparse==0.3.0
urllib3==1.25.3
virtualenv==16.7.5
virtualenv-clone==0.5.3
Werkzeug==0.16.0
